# server/src/api-docs/openapi.yml
openapi: 3.1.0
info:
  title: API de Autenticación y Usuario
  version: "1.0.0"
  description: |
    Especificación OpenAPI generada a partir de la tabla provista. Se preservan los
    mensajes, restricciones y flujos descritos (usuarios parciales/completos, cookies
    de sesión y de recuperación, verificación por token, etc.).

    Notas:
    - Donde se indica “debe pertenecer a un dominio permitido”, este documento no
      impone un patrón específico; queda a implementación (véase `x-allowedDomains`).
    - Para **usuarios parciales** y **completos**, se reflejan estados en `User.status`.
servers:
  - url: https://api.ejemplo.com
    description: Servidor principal
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Cookie de sesión. Requerida donde se indica “Requiere autenticación”.
    recoveryAuth:
      type: apiKey
      in: cookie
      name: pw_recovery
      description: Cookie de recuperación (emitida en `/password-recovery/verify`).
  parameters:
    emailQuery:
      name: email
      in: query
      required: true
      schema:
        $ref: '#/components/schemas/Email'
      description: Email válido; **debe pertenecer a un dominio permitido**.
      x-allowedDomains: []  # completar en implementación
    tokenQuery:
      name: token
      in: query
      required: true
      schema:
        type: string
        format: uuid
      description: Token de verificación (UUID).
...
  schemas:
    Email:
      type: string
      format: email
      description: |
        Email válido; **debe pertenecer a un dominio permitido** (la lista exacta queda a implementación).
      examples: [ "persona@ejemplo.com" ]
    Password:
      type: string
      description: |
        Contraseña de 8–32 caracteres; **debe incluir** mayúscula, minúscula, número y carácter especial.
      minLength: 8
      maxLength: 32
      pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z0-9]).{8,32}$'
      examples: [ "Abcdef1!" ]
...
    ErrorResponse:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details:
          description: Información adicional opcional del error.
          nullable: true
      required: [ code, message ]
paths:
...
  /password-recovery:
    get:
      tags: [ Recuperación de contraseña ]
      summary: Envía email para recuperar contraseña
      security: []  # sin autenticación
      parameters:
        - $ref: '#/components/parameters/emailQuery'
      responses:
        '200':
          description: |
            (1) Email de recuperación enviado, (2) Usuario no encontrado (respuesta neutral).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Email de recuperación enviado"
                      - "Si el usuario existe, se envió el email de recuperación"
              examples:
                sent: { value: { message: "Email de recuperación enviado" } }
                neutral: { value: { message: "Si el usuario existe, se envió el email de recuperación" } }
        '400':
          description: Falta email.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                missing_email: { value: { code: "bad_request", message: "Falta email" } }
        '500':
          description: Error inesperado al enviar correo.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                mail_error: { value: { code: "internal_error", message: "Error inesperado al enviar correo" } }
    put:
      tags: [ Recuperación de contraseña ]
      summary: Cambia la contraseña
      description: Requiere **password recovery cookie** (emitida por `/password-recovery/verify`).
      security:
        - recoveryAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ password ]
              properties:
                password: { $ref: '#/components/schemas/Password' }
      responses:
        '200':
          description: Contraseña cambiada con éxito.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: "Contraseña cambiada con éxito" }
        '400':
          description: |
            (1) No se proporcionó contraseña, (2) Formato de contraseña incorrecto,
            (3) Falta cookie de recuperación.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                password_missing: { value: { code: "bad_request", message: "No se proporcionó contraseña" } }
                password_format:  { value: { code: "bad_request", message: "Formato de contraseña incorrecto" } }
                recovery_cookie:  { value: { code: "bad_request", message: "Falta cookie de recuperación" } }
  /password-recovery/verify:
    get:
      tags: [ Recuperación de contraseña ]
      summary: Verifica token para habilitar cambio de password
      description: Requiere **token válido**.
      security: []  # sin autenticación adicional; validación vía query
      parameters:
        - $ref: '#/components/parameters/emailQuery'
        - $ref: '#/components/parameters/tokenQuery'
      responses:
        '200':
          description: |
            (1) Usuario no encontrado (respuesta neutral),
            (2) Verificación exitosa; continúa flujo (emite cookie de recuperación).
          headers:
            Set-Cookie:
              description: Puede establecer cookie de recuperación.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    enum:
                      - "Usuario no encontrado"
                      - "Verificación exitosa"
              examples:
                neutral: { value: { message: "Usuario no encontrado" } }
                ok:      { value: { message: "Verificación exitosa" } }
        '400':
          description: Falta email o token / error de validación del esquema.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                missing_params: { value: { code: "bad_request", message: "Falta email o token" } }
                schema_error:   { value: { code: "bad_request", message: "Error de validación del esquema" } }
        '401':
          description: Error de verificación.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              examples:
                verify_error: { value: { code: "unauthorized", message: "Error de verificación" } }
...